+++ packages/nestjs-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @feedback-forge/nestjs-plugin
 
+## 0.3.0-beta.1
+
+### Patch Changes
+
+- Angular Enhancements, Express direct, documentation updates
+- Updated dependencies
+  - @feedback-forge/plugin-common@0.3.0-beta.1
+
 ## 0.3.0-beta.0
 
 ### Minor Changes