+++ packages/payload-plugin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @feedback-forge/payload-plugin
 
+## 0.3.0-beta.1
+
+### Patch Changes
+
+- Angular Enhancements, Express direct, documentation updates
+- Updated dependencies
+  - @feedback-forge/integration-github@0.3.0-beta.1
+  - @feedback-forge/core@0.3.0-beta.1
+  - @feedback-forge/integration-jules@0.3.0-beta.1
+  - @feedback-forge/react-widget@0.3.0-beta.1
+
 ## 0.3.0-beta.0
 
 ### Minor Changes