+++ packages/react-widget/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @feedback-forge/react-widget
 
+## 0.3.0-beta.1
+
+### Patch Changes
+
+- Angular Enhancements, Express direct, documentation updates
+
 ## 0.3.0-beta.0
 
 ### Minor Changes